requirements.txt
pandas==1.1.3
ffmpeg_python==0.2.0
torch==1.6.0
opencv_python_headless==4.1.1.26
tqdm==4.51.0
torchvision==0.7.0
addict==2.2.1
numpy==1.18.1
ffmpeg==1.4
Pillow==8.0.1
PyYAML==5.3.1
